change to long counter

+++ 0x04-pointers_arrays_strings/7-puts_half.c
@@ -5,7 +5,7 @@
  */
 void puts_half(char *str)
 {
-	int size;
+	long int size;
 
 	size = _strlen(str);
 	size = (size % 2 == 0) ? size / 2 : (size - 1) / 2;
@@ -22,9 +22,9 @@ void puts_half(char *str)
  * @s: string to becalculated
  * Return: size of string
  */
-int _strlen(char *s)
+long int _strlen(char *s)
 {
-	int cont = 0;
+	long int cont = 0;
 
 	while (*(s + cont) != '\0')
 		cont++;
